# perf(spanner): improve mutationProto allocations and performance