Fixes: https://github.com/googleapis/google-cloud-go/issues/12680