@rahul2393 2025-08-20
The benchmark results indicate a significant performance improvement. The mutationsProto function is now approximately 81% faster, uses 57% less memory, and makes 16% fewer memory allocations on average.

```
goos: darwin
goarch: arm64
pkg: cloud.google.com/go/spanner
cpu: Apple M1 Pro
                                            │  before.txt  │              after.txt              │
                                            │    sec/op    │   sec/op     vs base                │
MutationsProto/small_number_of_mutations-10   13.778µ ± 1%   1.548µ ± 3%  -88.77% (p=0.000 n=20)
MutationsProto/large_number_of_mutations-10    43.83µ ± 1%   30.51µ ± 1%  -30.38% (p=0.000 n=20)
MutationsProto/mixed_type_of_mutations-10     13.289µ ± 1%   1.140µ ± 1%  -91.42% (p=0.000 n=20)
geomean                                        20.02µ        3.776µ       -81.14%

                                            │  before.txt  │              after.txt               │
                                            │     B/op     │     B/op      vs base                │
MutationsProto/small_number_of_mutations-10   8.844Ki ± 0%   2.961Ki ± 0%  -66.52% (p=0.000 n=20)
MutationsProto/large_number_of_mutations-10   67.10Ki ± 0%   59.16Ki ± 0%  -11.84% (p=0.000 n=20)
MutationsProto/mixed_type_of_mutations-10     7.979Ki ± 0%   2.164Ki ± 0%  -72.88% (p=0.000 n=20)
geomean                                       16.79Ki        7.237Ki       -56.90%

                                            │ before.txt  │              after.txt              │
                                            │  allocs/op  │  allocs/op   vs base                │
MutationsProto/small_number_of_mutations-10    84.00 ± 0%    67.00 ± 0%  -20.24% (p=0.000 n=20)
MutationsProto/large_number_of_mutations-10   1.343k ± 0%   1.321k ± 0%   -1.64% (p=0.000 n=20)
MutationsProto/mixed_type_of_mutations-10      64.00 ± 0%    49.00 ± 0%  -23.44% (p=0.000 n=20)
geomean                                        193.3         163.1       -15.63%
```